requirements.txt
paho-mqtt
scikit-learn
numpy
xgboost
flask
